requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
langchain==0.0.335
langchain-openai==0.0.2
langchain-core==0.1.4
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1 